[blas] Optimize gemv and gemm for float16 data type
This pull request aims to optimize the performance of gemv and gemm operations when the data type is float16.
This is achieved by converting the half-precision data to single-precision and utilizing the cblas functions instead of manually computing them using for loops.
This change should improve the overall efficiency of these operations without requiring ARM NEON support.

** Self-evaluation**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>